+++ NEWS.md
@@ -10,6 +10,7 @@ Bugs:
 
 * `save_tt("file.pdf")` works with colors. Thanks to @olivedv for the report and solution #395.
 * `group_tt(i=vec)`: `vec` can be a factor vector
+* `style_tt(align="d")` with empty strings (`modelsummary::datasummary_balance()` test)
 
 New: